google-meet-transcript-summarizer/requirements.txt
requests
python-dotenv
sounddevice
numpy
playwright
scipy
faster-whisper